Move WordQuiz related filed to new folder
